+++ lib/pages/palabra-single-screen.dart
@@ -2,6 +2,7 @@ import 'package:clipboard_manager/clipboard_manager.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
 
+import 'package:audioplayers/audioplayers.dart';
 import 'package:vocabulary_builder/model/main.dart';
 import 'package:vocabulary_builder/model/palabra.model.dart';
 import 'package:vocabulary_builder/utils/colors.dart';
@@ -13,6 +14,8 @@ import 'package:vocabulary_builder/widgets/ui/text/row-item.dart';
 import 'package:vocabulary_builder/widgets/palabra/single-palabra-card.dart';
 import 'package:flutter_web_browser/flutter_web_browser.dart';
 import 'package:scoped_model/scoped_model.dart';
+
+import 'dart:async';
 
 class SinglePalabraScreen extends StatelessWidget  {
   
@@ -440,7 +443,13 @@ class SinglePalabraScreen extends StatelessWidget  {
       height: 40.0,
       width: length < 8 ? length * 15.0 : length * 11.0,
       child: InkWell(
-          onTap: () => model.speak(text),
+          onTap: () => reproduceAudio(
+            model.userLang == 'en' ? 'es' : 'en',
+            model.userLang == 'en' 
+            ? _palabra.traduccion
+            : _palabra.palabra, 
+            text
+          ),
           onLongPress: () => _copy(text, context),
           child: Center(
             child: Text(
@@ -454,6 +463,17 @@ class SinglePalabraScreen extends StatelessWidget  {
         ),
       )
     );
+  }
+
+  Future<Null> reproduceAudio(String lang, String word, String word2) async {
+    
+    final AudioPlayer audioPlayer = AudioPlayer();
+    final String curatedString = word2.replaceAll(new RegExp(r"\s+\b|\b\s"), "");
+    final String url = 'https://s3.amazonaws.com/vocabulary-polly-sound-files/$lang-$word-$curatedString.mp3';
+
+    audioPlayer
+    .play(url)
+    .catchError((error) => print(error));
   }
 
   void _copy(String text, BuildContext context) {